requirements.txt
Django==1.11.11
PyMySQL==0.9.3
djangorestframework==3.10.3
